requirements.txt
pyyaml >= 3.12
discord.py >= 2.3.2
reqests >= 2.31.0